Added JavaFX Login Screen, started a Patient Portal, and implemented switching Scenes
I added a Login screen using FXML (it is not functional yet) and controllers. Using these controllers, I made it to where when you press the Patient Login button, it will take you to a work-in-progress Patient Portal.

What's left to be done is to add all the other screens, add the functionalities to the buttons, add the database interactions, and to make it look nice.

That's only 4 things when you think about it.