Added Create Account screen, updated Portal, added List selection
I have added a basic Create Account screen (fields are non-functional) and updated the Patient Portal to be a bit more complete. A list that is intended to show the patient's records has been added, as well as the ability to select it and view it in a textbox. Message screen is not yet added.

I plan on eventually modifying the controller classes to have a Constructor that allows for the passing of a userID parameter. This allows us to stay logged in across screens by just passing the userID from login/account creation to each following screen (except for the login screen when logging out!)